requirements.txt
requests 
beautifulsoup4 
google-generativeai 
ipython
streamlit